finalisation du projet

+++ controller/BackendController.php
@@ -3,43 +3,79 @@
 require_once('dao/ArticleDAO.php');
 require_once('dao/CommentDAO.php');
 require_once('dao/UserDAO.php');
+require_once "service/ConnectionService.php";
 
 class BackendController
 {
     private $articleDAO;
     private $commentDAO;
     private $userDAO;
+    private $connectionService;
 
     public function __construct()
     {
         $this->articleDAO = new ArticleDAO();
         $this->commentDAO = new CommentDAO();
         $this->userDAO = new UserDAO();
+        $this->connectionService = new ConnectionService();
     }
 
     public function displayAdminHome()
     {
-        $articles = $this->articleDAO->getArticles();
-        require_once('view/backend/admin/admin_home.php');
+        if ($this->connectionService->isConnected()) {
+            $articles = $this->articleDAO->getArticles();
+            $articlesHeader = $this->articleDAO->getArticles();
+            $articlesHeader = $this->articleDAO->getArticles();
+            $isConnected = true;
+            require_once('view/backend/admin_home.php');
+        } else {
+            header('Location: ?action=login');
+        }
     }
 
     public function modifyArticle($articleId)
     {
-        require_once('view/backend/admin/modify_article.php');
+        if ($this->connectionService->isConnected()) {
+            $articlesHeader = $this->articleDAO->getArticles();
+            $article = $this->articleDAO->getArticleById($articleId);
+            $isConnected = true;
+            require_once('view/backend/modify_article.php');
+        } else {
+            header('Location: ?action=login');
+        }
+
     }
 
     public function deleteArticle($articleId)
     {
-        $this->articleDAO->delete($articleId);
-        header('Location:?action=adminHome');
+        if ($this->connectionService->isConnected()) {
+            $this->commentDAO->deleteAllFromArticleId($articleId);
+            $this->articleDAO->delete($articleId);
+            header('Location:?action=adminHome');
+        } else {
+            header('Location: ?action=login');
+        }
     }
 
-    public function displayAddArticle($titlearticles, $content)
+    public function displayAddArticle()
     {
-        $articles = $this->articleDAO->postArticle($titlearticles, $content);
-       require_once('view/backend/admin/postArticle.php');
-       $articles = $this->articleDAO->modify_article($titlearticles, $content);
-       require_once('view/backend/admin/modify_article.php');
+        if ($this->connectionService->isConnected()) {
+            $articlesHeader = $this->articleDAO->getArticles();
+            $isConnected = true;
+            require_once('view/backend/post-article.php');
+        } else {
+            header('Location: ?action=login');
+        }
+    }
+
+    public function postArticle($title, $content, $pseudonyme)
+    {
+        if ($this->connectionService->isConnected()) {
+            $this->articleDAO->create($title, $content, $pseudonyme);
+            header('Location: ?action=adminHome');
+        } else {
+            header('Location: ?action=login');
+        }
     }
 
     public function displayLoginPage()
@@ -47,11 +83,65 @@ class BackendController
         require_once('view/backend/login.php');
     }
 
+    public function showReportedComments($articleId)
+    {
+        if ($this->connectionService->isConnected()) {
+            $articlesHeader = $this->articleDAO->getArticles();
+            $comments = $this->commentDAO->getReportedComments($articleId);
+            $article = $this->articleDAO->getArticleById($articleId);
+            $isConnected = true;
+            require_once "view/backend/comment_manager.php";
+        } else {
+            header('Location: ?action=login');
+        }
+    }
+
+    public function deleteComment($commentId)
+    {
+        if ($this->connectionService->isConnected()) {
+            $comment = $this->commentDAO->getById($commentId);
+            $this->commentDAO->delete($commentId);
+            header('Location: ?action=showReportedComments&articleId=' . $comment->getArticleId());
+        } else {
+            header('Location: ?action=login');
+        }
+    }
+
+    public function modifyArticleAction($articleId, $title, $content)
+    {
+        if ($this->connectionService->isConnected()) {
+            $this->articleDAO->update($articleId, $title, $content);
+            header('Location: ?action=detailArticle&articleId=' . $articleId);
+        } else {
+            header('Location: ?action=login');
+        }
+    }
+
+    public function login()
+    {
+        if ($this->connectionService->isConnected()) {
+            header('Location: ?action=adminHome');
+        } else {
+            $articlesHeader = $this->articleDAO->getArticles();
+            require_once "view/backend/page_connexion.php";
+        }
+    }
+
     public function loginAction($pseudo, $password)
     {
-        //TODO: Créer une méthode dans la DAO pour récupérer l'utilisateur qui a le bon pseudo
-        //TODO: Comparer si le mot de passe saisit est le bon
-        //TODO: Si OK -> redirige l'utilisateur vers l'espace admin // Si KO -> redirige vers la page de login
+        $user = $this->userDAO->getByPseudonyme($pseudo);
+        if (password_verify($password, $user->getPassword())) {
+            $this->connectionService->initConnection();
+            header('Location: ?action=adminHome');
+        } else {
+            header('Location: ?action=login');
+        }
+    }
+
+    public function logout()
+    {
+        $this->connectionService->logout();
+        header('Location: ?action=login');
     }
 
 }
